pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-compare"
version = "2.0.0"
description = "A Python tool to compare responses from multiple Large Language Models (LLMs)"
readme = "README.md"
requires-python = ">=3.9,<4.0"
license = {text = "MIT"}
authors = [
    {name = "LLM Compare Contributors"}
]
keywords = ["llm", "ai", "comparison", "litellm", "gpt", "claude", "gemini"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "litellm>=1.77.1,<2.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "aiohttp>=3.9.0,<4.0.0",
    "rich>=13.7.0,<14.0.0",
    "pyyaml>=6.0.1,<7.0.0",
    "tenacity>=8.2.3,<9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.9",
    "mypy>=1.7.1",
    "types-pyyaml>=6.0.12",
]

[project.scripts]
llm-compare = "llm_compare.cli:main"

[project.urls]
Homepage = "https://github.com/ParadoxReagent/llm_test"
Repository = "https://github.com/ParadoxReagent/llm_test"
Issues = "https://github.com/ParadoxReagent/llm_test/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["llm_compare*"]
exclude = ["tests*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--cov=llm_compare",
    "--cov-report=term-missing",
    "--cov-report=html",
]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by formatter
    "B008", # do not perform function calls in argument defaults
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_calls = false
warn_redundant_casts = true
warn_unused_ignores = true
